requirements.txt
boto3
flask
PyYAML==6.0
